comms: Cleanup connection managment to allow for use with different connection modes
So far we were only able to connect to the DE1 via BLE. To be able to remove this
restriction we need not only to safe the connectivity mode as well as the but build
generic versions of connection managment functions.
In this commit we also add the ability to search for USB devices and connect to them.

Signed-off-by: Mimoja <git@mimoja.de>